requirements.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
decorator==5.1.1
dj-database-url==0.5.0
Django==3.2.13
django-environ==0.8.1
django-heroku==0.3.1
djangorestframework==3.13.1
executing==0.8.3
factory-boy==3.2.1
Faker==13.7.0
flake8==4.0.1
gunicorn==20.1.0
idna==3.3
ipython==8.3.0
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.12.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.2.1.post0
urllib3==1.26.9
wcwidth==0.2.5
whitenoise==6.1.0
